google_cloud_functions/dailysql/requirements.txt
google-cloud-bigquery==1.24.0